Added tests for util

+++ jnvi-util/src/main/java/uk/ac/hud/jnvi/memory/DirectLong.java
@@ -27,7 +27,7 @@ public class DirectLong extends DirectMemory {
 	}
 
 	public void set(long index, long value) {
-		UnsafeAccessor.putFloat(getOffsetAddress(index), value);
+		UnsafeAccessor.putLong(getOffsetAddress(index), value);
 	}
 
 	public long get(long index) {
